+++ changelog.md
@@ -3,6 +3,7 @@
 - 2023-10-28
   - Added a new demo for calculating contour plots.
   - Added the `utils/Params` class for retrieving URL params by their proper type.
+  - Added/Refactored helper function `utils/clearDuplicateVertices`.
 - 2023-10-07 [v1.21.0]
   - Adding the `BezierPath.fromCurve(CubicBezierCurve)` static function.
   - Adding the extended demo: Trimming Bézier curves.